# app/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "eth-tools API",
    "description": "Self-maintaining runtime for ERC-8004 trustless agents.",
    "license": {
      "name": "MIT"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "https://eth-tools.dev",
      "description": "Production"
    },
    {
      "url": "https://preview.eth-tools.dev",
      "description": "Preview"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local dev"
    }
  ],
  "paths": {
...
    "/api/v1/reputation/give": {
      "post": {
        "tags": [
          "reputation"
        ],
        "operationId": "reputation_give",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReputationGiveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Submitted reputation feedback",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "400": {
            "description": "bad input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "missing bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "402": {
            "description": "x402 payment required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "wallet rails not yet wired (feat/phase-6-wallet-rails)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reputation/{chain}/{agent_id}": {
      "get": {
        "tags": [
          "reputation"
        ],
        "operationId": "reputation_read",
        "parameters": [
          {
            "name": "chain",
            "in": "path",
            "description": "Chain name or chain_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "agent_id",
            "in": "path",
            "description": "uint256 as decimal string",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List feedback rows for an agent, newest first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReputationList"
                }
              }
            }
          },
          "400": {
            "description": "bad input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "chain not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/validation/request": {
      "post": {
        "tags": [
          "validation"
        ],
        "operationId": "validation_request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Validation request submitted on-chain",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "400": {
            "description": "bad input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "missing bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "402": {
            "description": "x402 payment required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "wallet rails not yet wired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/validation/respond": {
      "post": {
        "tags": [
          "validation"
        ],
        "operationId": "validation_respond",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidationRespond"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Validation response submitted on-chain",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "400": {
            "description": "bad input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "missing bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "402": {
            "description": "x402 payment required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "wallet rails not yet wired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/validation/{chain}/{request_hash}": {
      "get": {
        "tags": [
          "validation"
        ],
        "operationId": "validation_read",
        "parameters": [
          {
            "name": "chain",
            "in": "path",
            "description": "Chain name or chain_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "request_hash",
            "in": "path",
            "description": "0x-prefixed 32-byte hex",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Single validation row",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationDetail"
                }
              }
            }
          },
          "400": {
            "description": "bad input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "chain or validation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApiErrorBody": {
        "type": "object",
        "description": "Top-level error envelope returned by every non-2xx response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorPayload"
          }
        }
      },
      "ApiErrorPayload": {
        "type": "object",
        "required": [
          "code",
          "policy_version",
          "evaluator"
        ],
        "properties": {
          "code": {
            "type": "string",
            "example": "AGENT_NOT_FOUND"
          },
          "evaluator": {
            "type": "string",
            "example": "api.lookup"
          },
          "override_hint": {
            "type": [
              "string",
              "null"
            ]
          },
          "policy_version": {
            "type": "string",
            "example": "v1"
          }
        }
      },
...
      "FeedbackDto": {
        "type": "object",
        "required": [
          "chain_id",
          "agent_id",
          "client_address",
          "feedback_index",
          "value",
          "value_decimals",
          "is_revoked",
          "tx_hash",
          "block_number"
        ],
        "properties": {
          "agent_id": {
            "type": "string"
          },
          "block_number": {
            "type": "integer",
            "format": "int64"
          },
          "chain_id": {
            "type": "integer",
            "format": "int64"
          },
          "client_address": {
            "type": "string"
          },
          "endpoint": {
            "type": [
              "string",
              "null"
            ]
          },
          "feedback_hash": {
            "type": [
              "string",
              "null"
            ]
          },
          "feedback_index": {
            "type": "integer",
            "format": "int64"
          },
          "feedback_uri": {
            "type": [
              "string",
              "null"
            ]
          },
          "is_revoked": {
            "type": "boolean"
          },
          "tag1": {
            "type": [
              "string",
              "null"
            ]
          },
          "tag2": {
            "type": [
              "string",
              "null"
            ]
          },
          "tx_hash": {
            "type": "string"
          },
          "value": {
            "type": "string",
            "description": "Decimal string (NUMERIC(40,0)). Combine with `value_decimals` to\ninterpret as a fixed-point score."
          },
          "value_decimals": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ReputationGiveRequest": {
        "type": "object",
        "required": [
          "chain",
          "agent_id",
          "value",
          "value_decimals"
        ],
        "properties": {
          "agent_id": {
            "type": "string"
          },
          "chain": {
            "type": "string"
          },
          "endpoint": {
            "type": [
              "string",
              "null"
            ]
          },
          "feedback_hash": {
            "type": [
              "string",
              "null"
            ],
            "description": "0x-prefixed 32-byte hex."
          },
          "feedback_uri": {
            "type": [
              "string",
              "null"
            ]
          },
          "tag1": {
            "type": [
              "string",
              "null"
            ]
          },
          "tag2": {
            "type": [
              "string",
              "null"
            ]
          },
          "value": {
            "type": "string",
            "description": "Decimal string for NUMERIC(40,0)."
          },
          "value_decimals": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ReputationList": {
        "type": "object",
        "description": "Concrete instantiation for stable schema name — see `AgentList` doc.",
        "required": [
          "data",
          "staleness_ms",
          "source"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeedbackDto"
            }
          },
          "source": {
            "type": "string",
            "example": "db"
          },
          "staleness_ms": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ValidationDetail": {
        "type": "object",
        "required": [
          "data",
          "staleness_ms",
          "source"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ValidationDto"
          },
          "source": {
            "type": "string",
            "example": "db"
          },
          "staleness_ms": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ValidationDto": {
        "type": "object",
        "required": [
          "chain_id",
          "request_hash",
          "validator_addr",
          "agent_id",
          "request_uri",
          "last_update"
        ],
        "properties": {
          "agent_id": {
            "type": "string"
          },
          "chain_id": {
            "type": "integer",
            "format": "int64"
          },
          "last_update": {
            "type": "string",
            "format": "date-time"
          },
          "request_hash": {
            "type": "string",
            "description": "0x-prefixed 32-byte hex."
          },
          "request_uri": {
            "type": "string"
          },
          "response": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "0..100 score; `None` until the validator responds."
          },
          "response_hash": {
            "type": [
              "string",
              "null"
            ]
          },
          "response_uri": {
            "type": [
              "string",
              "null"
            ]
          },
          "tag": {
            "type": [
              "string",
              "null"
            ]
          },
          "validator_addr": {
            "type": "string",
            "description": "0x-prefixed 20-byte hex."
          }
        }
      },
      "ValidationRequest": {
        "type": "object",
        "required": [
          "chain",
          "agent_id",
          "validator_addr",
          "request_uri",
          "request_hash"
        ],
        "properties": {
          "agent_id": {
            "type": "string"
          },
          "chain": {
            "type": "string"
          },
          "request_hash": {
            "type": "string",
            "description": "0x-prefixed 32-byte hex (the canonical hash of the request_uri body)."
          },
          "request_uri": {
            "type": "string"
          },
          "tag": {
            "type": [
              "string",
              "null"
            ]
          },
          "validator_addr": {
            "type": "string",
            "description": "0x-prefixed 20-byte hex."
          }
        }
      },
      "ValidationRespond": {
        "type": "object",
        "required": [
          "chain",
          "request_hash",
          "response"
        ],
        "properties": {
          "chain": {
            "type": "string"
          },
          "request_hash": {
            "type": "string",
            "description": "0x-prefixed 32-byte hex of the request being responded to."
          },
          "response": {
            "type": "integer",
            "format": "int32",
            "description": "0..100 score."
          },
          "response_hash": {
            "type": [
              "string",
              "null"
            ],
            "description": "0x-prefixed 32-byte hex of the response body."
          },
          "response_uri": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      }
    }
  },
...
}